+++ UVtools.Core/Objects/StaticObjects.cs
@@ -5,7 +5,9 @@ namespace UVtools.Core.Objects
 {
     public static class StaticObjects
     {
-        public static SHA256 Sha256 { get; } = SHA256.Create();
+        public static readonly SHA256 Sha256 = SHA256.Create();
+
+        public static readonly string[] LineBreakCharacters = {"\r\n", "\r", "\n"};
 
         // Compute the file's hash.
         public static byte[] GetHashSha256(string filename)